Add symlink copy with `followSymlinks` option
The current behavior is always to resolve symbolic links to actual
files.

This change introduces a `followSymlinks` option to `src()` (default:
`true`). When the option is set to `false`, then each encountered
symlink is stored in `file.symlink`. When `dest()` is invoked, the
`symlink` value is used to reconstruct an identical symlink.

It is important to note that the symbolic link path remains completely
unmodified, so care should be taken when sourcing symlinks with absolute
paths.

This patch also includes related unit tests